chore: load project settings from teamsapp.yml

+++ packages/fx-core/src/core/middleware/projectSettingsWriter.ts
@@ -41,13 +41,16 @@ export const ProjectSettingsWriterMW: Middleware = async (
     if (projectSettings === undefined) return;
     try {
       if (isV3Enabled()) {
-        const settings: Settings = {
-          trackingId: projectSettings.projectId,
-          version: projectSettings.version!,
-        };
+        // TODO: remove below logic when folder structure change finished
         const settingFile = getProjectSettingsPath(inputs.projectPath);
-        await fs.writeFile(settingFile, JSON.stringify(settings, null, 4));
-        TOOLS?.logProvider.debug(`[core] persist project setting file: ${settingFile}`);
+        if (await fs.pathExists(settingFile)) {
+          const settings: Settings = {
+            trackingId: projectSettings.projectId,
+            version: projectSettings.version!,
+          };
+          await fs.writeFile(settingFile, JSON.stringify(settings, null, 4));
+          TOOLS?.logProvider.debug(`[core] persist project setting file: ${settingFile}`);
+        }
       } else {
         projectSettings = convertProjectSettingsV3ToV2(projectSettings as ProjectSettingsV3);
         const solutionSettings = projectSettings.solutionSettings;
